+++ CHANGELOG.md
@@ -1,5 +1,19 @@
 # Changelog
 
+## 1.4.0 - TBA
+
+- Add theme class support, see #260
+- Fix on iOS, see #101
+
+## 1.3.2 - 2015-02-17
+
+- Add options to disable days, see #171
+- Add SCSS file, see #130
+
+## 1.3.1 - 2015-02-04
+
+- Fix for bug introduced in #166; fixes #224, #230
+
 ## 1.3.0 - 2015-01-09
 
 - Fix #79 moving through months can cause calendar to "bounce around"